refactor: API Key 타입 import 경로 수정
