+++ apps/client/src/widgets/home/model/useCreateApiKey.ts
@@ -2,7 +2,7 @@ import { authQueryKeys, authUrl, post } from '@repo/shared/api';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-import { ApiKeyResponse } from './types';
+import { ApiKeyResponse } from '@/entities/home';
 
 export const useCreateApiKey = (
   options?: Omit<
